backend/requirements.txt
fastapi[standard]==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
alembic==1.14.*
google-generativeai==0.8.*
pydantic-settings==2.7.*
python-jose[cryptography]==3.3.*
aiosmtplib==3.0.*
python-multipart==0.0.*
email-validator==2.2.*
